Change in how genes are configured
We now do not connect genes on transcript rafts. Rather, each gene that connects with a transcript, gets that transcript added to their repertoire. Transcripts that are not connected to genes will still build. Current concern is for transcript rafts to still be huge (even without crossing gene locations).
However, this will not effect the speed of the code too much (hopefully).